refac: centralize xml parsing orchestration
